+++ contracts/cw20-base/src/multitest/receiver.rs
@@ -1,5 +1,7 @@
 use cosmwasm_std::{Binary, DepsMut, Env, MessageInfo, Response, StdError, Uint128};
-use sylvia::interface;
+use sylvia::{contract, interface};
+
+use super::receiver_contract::ReceiverContract;
 
 #[interface]
 pub trait Receiver {
@@ -13,4 +15,19 @@ pub trait Receiver {
         amount: Uint128,
         msg: Binary,
     ) -> Result<Response, Self::Error>;
+}
+
+#[contract]
+impl Receiver for ReceiverContract {
+    type Error = StdError;
+
+    fn receive(
+        &self,
+        _ctx: (DepsMut, Env, MessageInfo),
+        _sender: String,
+        _amount: cosmwasm_std::Uint128,
+        _msg: cosmwasm_std::Binary,
+    ) -> Result<Response, Self::Error> {
+        Ok(Response::default())
+    }
 }
